Issue 7958: 1.Added versioning support prototype to database 2.Added versioning support to document meta
git-svn-id: https://svn.imcode.com/imcode/imcms/trunk@13188 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ src/main/java/com/imcode/imcms/api/I18nMeta.java
@@ -32,7 +32,7 @@ public class I18nMeta implements Serializable, Cloneable {
 	private Long id;
 	
 	@Column(name="meta_id")
-	private Integer metaId;
+	private Long metaId;
 	
 	@Column(name="meta_enabled")
 	private Boolean enabled;
@@ -120,11 +120,11 @@ public class I18nMeta implements Serializable, Cloneable {
 		this.id = id;
 	}
 
-	public Integer getMetaId() {
+	public Long getMetaId() {
 		return metaId;
 	}
 
-	public void setMetaId(Integer metaId) {
+	public void setMetaId(Long metaId) {
 		this.metaId = metaId;
 	}
